refactor(runtime): extract hand and a2a tool modules from tool_runner
- `hand` — list / activate / status / deactivate. Each verbatim move
  from the original tool_runner.rs body; only visibility upgraded to
  pub(super).
- `a2a` — discover and send for cross-instance agent communication.
  Pulls in the existing taint sinks (check_taint_outbound_text,
  check_taint_net_fetch) via super:: so the SECURITY (#3786) trust
  gate and exfil-data check keep their original ordering — taint
  rejection still precedes the trusted-agent allowlist check.

Behavior-preserving. The stale "// Location tool" section header
left behind in an earlier batch is intentionally not touched here;
it'll go when the fs/media helpers (resolve_file_path_ext et al.)
move out and the surrounding section comments are rebuilt.

mod.rs: 12,510 -> 12,360 LOC (-150). cargo check + cargo clippy --lib
both clean.

Refs #3710
